+++ test/shared/index.js
@@ -80,6 +80,34 @@ exports.test = function(name) {
       });
     });
 
+    it('should return a promise if no callback provided', function(done){
+      var path = 'test/fixtures/' + name + '/user.' + name;
+      var locals = { user: user };
+      var result = cons[name](path, locals);
+
+      result.then(function (html) {
+        html.should.equal('<p>Tobi</p>');
+        done();
+      })
+      .catch(function (err) {
+        done(err);
+      });
+    });
+
+    it('should return a promise if no callback provided (string)', function(done){
+      var str = fs.readFileSync('test/fixtures/' + name + '/user.' + name).toString();
+      var locals = { user: user };
+      var result = cons[name].render(str, locals);
+
+      result.then(function (html) {
+        html.should.equal('<p>Tobi</p>');
+        done();
+      })
+      .catch(function (err) {
+        done(err);
+      });
+    });
+
     it('should be exposed in the requires object', function(){
       var should = require('should'),
         requiredName;